requirements.txt

langchain-community
openai
langchain-openai
langchain-text-splitters
langchain-classic
langchain-core
chromadb
python-dotenv
slowapi
fastapi
pydantic
mssql-python
streamlit
pypdf
uvicorn
deepeval
pdfplumber
fastapi